Bash: removed CLICOLOR_FORCE=1 - was causing completions to break, etc.
Basically, any tool chain that relied on output from ls was breaking. This was causing completions to break because the bash_completions shell script sourced all the application specific completions by `ls`ing /usr/local/etc/bash_completion.d and iterating over the results, sourcing each. But with CLICOLOR_FORCE=1 set, ls included a bunch of escape characters, which fucked up the output and didn't allow the bash script to iterate over it.

I thought about creating a custom function/alias for ls, where we set that, for the times we do want to force colors. But after experimentation, I've find I'd never want to force colors. The original intent was to see the file listing with its original colorization after running through grep, but even this gets fucked up with colorization on: the escape sequences cause grep to not match things it should match, leading to incorrect output.