Update .bash_profile to be more robust & readable.
We now do more sanity checks. For example, before setting $PS1, and
color options, we now check if we're in an interactive shell. We also
check for the existence of programs (like homebrew and macvim) before
trying to use them.

The file is also better organized, with nice headers and logical
groupings of commands.

Finally, I modified $PS1/$PS2 to remove the \W escape sequence, due to
a bug in the current version of OS X Mavericks and homebrew-installed
BASH.