Git: Create wrapper to fix Kaleidoscope bug
There is a bug with the current version of Kaleidoscope and git where,
if you launch Kaleidoscope from the terminal using the ksdiff utility,
it will crash when trying to launch git as a subprocess. This doesn't
appear to have any real effect, but it causes annoying OS X error
messages to pop up.

This script acts as a wrapper to ksdiff. It first checks if
Kaleidoscope is open and, if not, will launch it from a new environment
which doesn't inherit settings from the current one (I believe the
error comes from certain variables being set in my highly customized
Bash environment, which Kaleidoscope can't deal with.) It then pauses a
second before running ksdiff as normal, which re-uses the existing
Kaleidoscope process, avoiding launching a new one tainted by this
terminal.

Hopefully I'll be able to delete this wrapper when Kaleidoscope fixes
the bug.