Bash: replace rsync alias with `mrsync` command.
Previously, we creates an alias for rsync that replaces `rsync` with
`rsync --a-bunch --of-options`. However, this wasn't great, because it
broke things like `rsync --daemon`.

We now have a command, ~/bin/mrsync, which does pretty much the same
thing (calls rsync with a sane set of defaults.) This leaves the
regular rsync command untouched, but keeps us from having to remember a
dozen command-line options.