Bash: replace bash aliases for `brew` with wrapper
Previously, we were unable to directly alias homebrew commands, because
homebrew uses a single monolithic `brew` command to run all
sub-commands. So if we wanted to modify the `brew linkapps` command to
use the '--local' option by default, we would have to create a new
alias like `brew-linkapps` instead of being able to alias `brew linkapps` directly.

This change adds the `brew-wrapper` script which allows us to look at
which subcommand and options are being passed to the `brew` app, modify
them, and call into the `brew` app directly.

We then alias this script as `brew` so that it invisibly handles this
task.