Vim: move Powerline init back to .vimrc.before
Previously, we had intended for PowerLine to only be used in GUI Vim sessions. Before commit b636e12, the init command was inside .vimrc.before, but guarded by a ‘if’ statement to check if the GUI was running. b636e12 moved the command to .gvimrc.before, without the ‘if’ guard. However, being in .gvimrc.before didn’t seem to work. Even running in GUI mode, PowerLine did not init.

I decided that PowerLine should run in both terminal and GUI mode. I had avoided it previously because you need a special font to support PowerLine. However, I run this font in Terminal.app anyway, so might as well run PowerLine there.

Now the terminal Vim looks quite nice, with the same 256-color color scheme as the GUI and the same PowerLine status line.