Vim: if opening >1 file, put each in their own tab
I first created a new section in my .vimrc.after file for running
commands after Vim has fully loaded.

To this, I added a function I created which checks the number of files
passed to Vim on startup. If there is more than one, we create a new
tab for each buffer.