Bash: create a set of helper functions for scripts
The non-executable file ~/bin/helpers.sh contains a collection of
functions which are useful for other functions in Bash scripts. A script
should source this file and then call the desired functions.

This file is not meant to be run directly by the user (and hence is not
executable.) However, if you really want to call these functions
directly, simply source the file in your Bash session and call them.

To start, there are three functions:

* get_script_name(): echoes the name of the script file the user called
(useful for printing help messages, etc.)

* get_script_path(): echoes the full path and filename of the called
script (useful if you want to write out a full command for another
script to later call you.)

* check_url(): uses cURL to check if we can retrieve the URL without
errors (used in hman, for example, to check if a man page exists on
Apple.com, instead of creating it ourselves.)