Bash: `hman` will try to open Apple.com's version
The `hman` command will check to see if the manpage is a standard
system manpage. If so, it will check Apple.com's website to see if the
manpage exists there. If it does, it opens that version in the browser.
If it does not exist, or the manpage is a non-standard one (for
example, from homebrew,) it continues to generate a local HTML version
of the manpage itself, as usual.