Bash: add echo_wrapped() to helpers.sh (plus todos)
This adds the function echo_wrapped() to ~/bin/helpers.sh. This
function takes a string and echoes it to stdout, word-wrapping the
output to the width of the terminal.

I also added a few 'TODO' comments to helpers.sh detailing future work.
The comments (obviously) shouldn't affect functionality at all.