Vim: update vim-session from 1.5 to 2.4.9
I forked vim-session a long time ago to add some features I felt were
missing. This was the repo that was listed as a submodule here.
However, the main vim-session repo almost as long ago pulled in my
changes, and has since added a bunch more, going from version 1.5 (or
something similar) to 2.4.9.

I changed the URL from the submodule to point to the main repo, thus
updating vim-session to the latest version. My fork is no longer
relevant and should not be used.