Bash: add 'sudo_gui' function to helpers.sh
This function executes a command with root privileges, like `sudo`.
However, unlike `sudo`, it prompts the user for their password via a
GUI dialog, not via the Terminal.

This is useful if your script is executing in the background, without a
Terminal session, and you need to elevate privileges. Without a
Terminal, you can't prompt the user to enter their password for `sudo`,
so this function comes in handy.