hman: save converted manpaged. Add -l & -r options
This is a significant rewrite of the structure of hman. It now supports
two different program arguments: -l to tell it to always open the local
version of the man page (rather than try to open standard OS X man
pages on Apple's site) and -r to tell it to force regeneration of
cached HTML versions of manpages.

The basic command to generate HTML output from a man page, and
the sed script that rewrites it into something somewhat sane,
are unchanged. However everything has been refactored to be contained
in functions.

We also now save the generated HTML files into
~/Library/Caches, and have logic for making sure these files aren't
stale. There is also logic to move generated supporting files (images,
etc., as seen in `hman troff`) to the right location, and rewriting the
HTML to reflect that location.