Vim: cleanup of .vimrc files. Margin in col 101.
In addition to removing redundant configuration and putting
configuration in the proper place (before, after, gvim, vim), there are
only two other changes of note.

The first is to place the highlighted margin column in line 101,
instead of 100. I want my lines to be 100 characters long or less. If
the color column is in column 100, then the 100th character would be
inside the column. At 101, I can write right up to that column, and I
know if I ever infringe on its space, I've gone over.

I also turned off showing the mode in the command bar. This was
redundant with Powerline modeline. This should clean up some visual
clutter and remove the "--INSERT--" message when going into insert
mode, etc.

One last minor change was getting rid of the old :Vimrc command, since
we replaced it with the much more robust :VimRC command.