.js: fix bug in GitHub TOC if first heading is h2
If the very first heading in the MarkDown content is an <h2>, before
there are any <h1> headings, there was previously a bug that prevented
the script from running. This patch fixes this.

Normally, we add <h2> elements as a new sub-list, contained in the
previous <li> element created by an <h1>. If there is no previous <h1>
element, then there is no <li> to add the sub-list to.

There was existing code that was supposed to check this condition, and
create an empty <li> to append the sub-list to, but it was checking if
the wrong element was null (`h1List`, which always exists; we should be
checking `h1li`.)