.js: GitHub TOC script now handles AJAX-y links
Many of GitHub's links aren't plain <a href="..."> links which point
the browser to a new URL, which it loads and executes. Instead, they
use a library called 'pjax' to fetch the new page via AJAX and replace
part of the page with this new content.

This caused a problem for out GitHub TOC because it only ran on page
load. So if a new MarkDown-containing page was loaded via pjax, we
woudn't add/update our TOC for it.

We fix this by creating a DOM Mutation Observer which watches the pjax
content area for new elements which may indicate the page has been
changed. We then re-run our 'add TOC' function.

It's hack-y in some ways, but it seems to work pretty well at the
moment.