.js: In GitHub TOC, handle styling on 0-1 h1 items
Our stylesheets now detect where there is only one or zero h1 (that is,
top-level) items in the table-of-contents, and styles them better.

In the case of 0 h1 items, it styles all the h2 items as if they were
h1 items.

In the case of 1 h1 item, its disables the text indent for the second
line of text (thus, instead of the second line of the h1 item being
indented, it is flush left.)