Vim: add modeline cmd updates existing modeline
The command to add a modeline to the current file (bound to <leader>ml)
will now search the file for an existing modeline. If one is found,
instead of adding another modeline, it will replace the existing
modeline with a new one (representing the current settings.)

If no existing modeline is found, one is added to the end of the file
as before. However, it will add a blank line between the last line of
the file and the modeline, if one wouldn't otherwise exist.

As a note, the existing modeline search regexes aren't quite as
sophisticated as Vim's native version, so it's possible that it may
generate false positives/negatives.