.js: In GitHub TOC, improve styling
The sizes are now adjusted to be slightly smaller, second-line indent
on multiline headings is gone, the line height is smaller, and the
margin between <li> is increased.

It's still not pretty, nor does it fit with the GH style very well. It
also looks especially bad if the window is narrow, since it hangs way
off to the left of the page. But it's usable.