Vim: switch numbers.vim to my own fork
numbers.vim has an incompatibility with Buffergator, where you can't
accurately tell it to ignore Buffergator's windows. I forked
numbers.vim on GitHub and fixed this incompatibility.

I have a pull request in with the upstream, main repo, but in the
meantime, I switched the submodule URL git pulls from to point to my
own fork so that we see the changes immediately.