Bash: fix hman bug to make output quieter
When you run `hman` on a man file, before it generates its HTML + supporting files, it tries to clean up any lingering files from the last time hman was run on that man file. I intended for the `rm` command that does this to have its stderr redirected to /dev/null, so that the user wouldn’t see an error message in their terminal if there were no support files to delete, but I used ‘>/dev/null’ instead of ‘2>/dev/null’. Oops.