Bash: `brew graph` opens PDF if no errors
In the brew-wrapper script, `brew graph` will now generate the PDF to a
unique location based off ~/Desktop/brew.pdf (instead of overwriting
that file if it already exists.) You can also pass in a filename as an
argument, and the PDF will be written to that file.

It then checks to see if GraphViz generated the PDF without warnings or
errors. If so, it opens up the PDF. If not, it will show you the
warning/error messages.