Vim: set textwidth to 100, & better formatoptions
We now set the textwidth to 100. However, we also set formatoptions
intelligently. Specifically, we don't include 't' as a part of it, so
Vim will not wrap lines of code at 100. However, we do include 'c', so
that it will wrap comments at 100. We also add some other
formatoptions, which are explained in the .vimrc.after, where we set
them.

Additionally, for gitcommit files, we *do* add 't' to formatoptions and
set the textwidth to 72, so that all lines in those files are, at most,
72 columns wide.