Bash: add (empty) ~/.bash_logout
This file is sourced by Bash before every logout action. Ubuntu's
default is to clear the screen, but I don't necessarily want to do
that, so I've just commented out the whole file. However, I added it to
the repo for reference, and as a reminder if we want to add future
functionality.