Vim: remove 'w', add 'a', to formatoptions
'w' added trailing whitespace to wrapped lines to indicate they
continued on the next line, but this goes against the 'no trailing
whitespace' code/commit message formatting rules.

'a' automatically updates the formatting of paragraphs as we insert or
delete text.

We also change the `set` command to use "+=" and "-=" instead of a
straight "=". Vim's help recommends this to support additional format
options that are added in the future.