Git: move filters to ~/.gitsupport/filters
The main "email" filters aren't in the repo because they contain
sensitive information (my real email address -- the filters filter it
out of any commits and replace it with public, sanitized versions.)

This commit changes the directory pointed to in ~/.gitconfig for the
filters. I then manually moved the filters into this directory.