Vim: properly set formatoptions
It looks like the "-=" expression only succeeds if every value on its
rhs is in the variable in the rhs. Thus, our line "-=tw2" would fail
because one or more of 't', 'w' or '2' was not in formatoptions.

This change manually removes each of those values from formatoptions
with individual "-=" statements. This should properly remove all from
formatoptions.

I don't *think* the same sort of bug applies to our "+=" expression.