Git: (template) set SourceTree to unwrap lines
In our user-defined template directory, we add a "sourcetreeconfig"
file, which is the file SourceTree uses to set per-repo config settings.

This config file has the effect of unwrapping word-wrapped line (for
example, those generated by our commit-msg hook.) It does this by
replacing non-double newlines with spaces -- i.e., if a paragraph is
separated from the previous one by a blank line, it leaves it be; if
not, it replaces the newline with a space.

Ideally, we'd set this option globally in SourceTree, but I don't see
an option for it.

This config allows us to have our cake and eat it too. In the Terminal
and on GitHub, commit messages are hard word-wrapped at 72 characters.
In SourceTree, they're soft-wrapped at the dialog boundary.

This may also affect messages of commits we make in ST (that is, the
messages with be modified when we write the commit.) However, this
should be an idempotent action, for properly formatted commit messages.