Git: change name of SourceTree replacement pref
In our user-defined template dir, we change the sourcetreeconfig file
to modify the name of our 'unwrap lines' setting. (As a reminder, this
setting unwraps hard-wrapped lines in commit messages in SourceTree.)
Specifically, we change the name from an auto-generated "Search text ->
Replacement Text", to the more descriptive "Unwrap Lines".

This should have no functional difference, except for what the config
is named in the SourceTree UI.