Bash: remove unused $color_prompt var from profile
As a vestige of the original Ubuntu ~/.profile, we detected the $TERM
type and then set the variable $color_prompt if we thought the term
supported colors.

The original Ubuntu .profile then took some later actions based off if
$color_prompt was set (like enabling a color $PS1.) However, in our
~/.bash_profile, we assume all terminals have color (because, come on,
who doesn't???) We therefore never actually used the $color_prompt
variable.

Accordingly, we remove the test and setting of $color_prompt for
~/.bash_profile.