Bash: tweak $PS1 if running in tmux
We tweak the way we set our $PS1 and $PS2 in ~/.bash_profile to support
using slightly different settings in tmux.

We first refactor the setting of $PS1 to extract the shell level
variable from being used directly in the $PS1 string (as $SHLVL), into
the variable $NUM_PROMPTS, which we set before setting $PS1, and use in
$PS1 in place of $SHLVL.

We then add code before we set $PS1 to detect if $TERM is tmux. If so,
we set $PROMPT_COLOR to be a slightly different color (same bg color,
but a light yellow fg color.)

If the $TERM is tmux, we also subtract 1 from $NUM_PROMPTS. This is
because tmux itself adds another level to $SHLVL, but seeing this
doesn't really tell us anything useful (that knowing we're running in
tmux, via the different prompt color and Powerline status line) doesn't
tell us.