Bash: add SSH and SCP shortcut for my VPS
Firstly, this requires that the MT_URL and MT_PORT environmental
variables are set before ~/.shell_commands is sourced. These values
contain potentially sensitive information, so they're kept out of source
control directly. Currently, they're set in ~/.shell_private.

In ~/.shell_commands, we add an alias, `mt.com`, which SSH'es into my
VPS.

In the same file, we also add a wrapper around `scp` which detects if
`mt.com` is one of the URLs and replaces it with the real VPS URL. If no
port was specified in the user's scp arguments, it also adds the port of
my VPS to the scp arguments. Thus, the command `scp mt.com:~/a_file.txt
./` will correctly download a_file.txt from my VPS, into the local dir.
(This is accomplished by writing a function which mangles the args to
replace the URL and/or port, then calls `scp` with those args.)