Vim: Add 'w' to format options
'w' causes Vim to assume lines with whitespace at the end continue a
paragraph. While we normally don't want whitespace at the end of lines,
maybe this will help with the "can't insert a newline into the middle
of a comment" bug (where, with 'a' in the formatoptions, Vim assumes
the two new lines are part of the same paragraph and immediately
autoformats them back into a single word-wrapped paragraph.)

Just to test it out. May revert later.