Vim: add command to toggle 'a' in formatoptions
`\a` is the keyboard shortcut in normal mode. On Vim startup, 'a' will
be in the formatoptions. Pressing \a will remove it; pressing \a again
will add it back again.

This is a bandaid fix for the problem of not being able to insert a
newline in the middle of a comment or other wrapped text. This is
because Vim tries to immediately autoformat the text, and (since there
isn't a blank line between the two lines, and Vim therefore thinks the
two lines are part of the some paragraph,) joins the two lines back
together, negating the newline we just entered.

With this change, we can toggle 'a' off, insert a newline and reformat
the text manually, then reenable 'a'.

Read the TODO in the comments for more of my thoughts on this.