Bash: set $OS based on current system
The first thing ~/.bash_profile will now do is detect the operating
system of the system it's running on, and set the globally exported
variable $OS to either "Mac" or "Linux".

This is part of our (still early) work to use the same dotfile branch
on both Mac and Linux. OS-specific configuration can now be guarded by
a `if [[ $OS == Mac ]]; then` (or 'Linux'), allowing for the same file
to be used on both systems without errors.