Bash: improve shell aliases
'rm' is now aliased to ask for confirmation when deleting >3 files, or
when using the '-r' option. It also verbosely prints its actions.

'make' will now use (# of CPUs) + 1 subprocesses when building projects
(this is the optimal number, according to tHe InTeRnEtZ), where the
exact number is determined dynamically by checking the number of CPUs
reported by the OS, then adding 1 to it, at login time.

Fixed a bug in 'xargs' where the replacement string '{}' wasn't being
set correctly.

'fg' (alias to `fgrep`) now ignore binary files.

Added more descriptive comments, so that each alias has a description of
all the options it turns on.