Bash: disable `keychain` wrapper around ssh, git
We used to wrap the `ssh` and `git` commands with a call to the
`keychain` utility, which would prompt us for our SSH key password (if
it wasn't already cached,) and then cache it for subsequent calls to
those commands.

We've now enabled ssh-agent forwarding on my Mac laptop (Son of
IveBook,) so that we don't use the Linux machine's SSH keys for
github.com anymore. Therefore, we've disabled the `keychain` wrapper for
now.

Not only does this eliminate a useless password prompt now (since we
rarely need to unlock the Linux machine's SSH keys,) but it also
prevents `keychain` from clobbering the SSH forwarding settings the Mac
sets up when it connects.