Bash: ~/.ssh/config instead of `mt.com` alias
We previously had created a complicated set of scripts around
connecting to my server, including an alias for SSHing to it (`mt.com`)
and a wrapper around `scp` which would insert the proper URI, port, and
username if it saw 'mt.com' as the URI being used.

However, it turns out we can eliminate all of this with ~/.ssh/config.
There, you can setup aliases, set their hostname, port, username and
even turn on ssh-agent forwarding (which we do for our server, so that
the server can use our local Mac OS X credentials when we're SSHed into
it.)

This is accomplishes the same thing our complicated scripting system
does, more elegantly and robustly, and works for both `ssh` and `scp`
(and `git` and anything else using SSH connections, most likely.)

Obviously ~/.ssh/config hasn't been checked into the repo because it
contains sensitive information (URIs, ports, usernames, etc.) but it
shouldn't be hard to recreate if need be. An example is:

Host mt.com
        HostName <URI>
        Port <Port>
        User <username>
        # Turn on ssh-agent forwarding
        ForwardAgent yes