Bash: `brew outdated` now runs `update` first
In my homebrew wrapper script, I modified the `brew outdated` command
to first run `brew update`, so that we have the freshest updates before
showing outdated formulae.

If the output is going to a terminal then we show the output of `brew
update`, followed by a divider, followed by the output of `brew
outdated`. If the output is instead being piped somewhere or similar,
then we just show the output of `brew outdated`.

The '-f'/'--fast' option to `brew outdated` (as in `brew outdated
--fast`) will cause `outdated` to skip running `update` first.