More robust arguments, help msg to `serve` command
The `serve` command runs a HTTP server, serving the current directory.
This commit adds a help message, accessible by passing '-h' or '--help'
as one of the arguments.

The arguments to `serve` are now more robust as well. Before, we had
two separate arguments for IP and port. Now, they are one argument in
standard form (for example, '0.0.0.0:8080').