Bash: add `gsudo` command, remove `sudo_gui`
Both commands do the same thing: run a terminal command with
administrator privileges, using a GUI to prompt the user for their
password (as opposed to regular `sudo`, which prompts the user for
their password in the terminal.)

This commit moves that functionality from a function in helpers.sh, to
its own script `gsudo`. This allows it to be used directly by the user,
or from scripts that don't source helpers.sh.

This commit also changes the code to escape quotes in the provided
terminal command, which would otherwise cause problems with AppleScript.