Bash: add tmux aliases `tnew` and `topen`
The `tnew` alias creates a new tmux session. It is called with one
argument, the name of the new session, as in `tnew my-session-name`.

The `topen` alias opens an existing session. It is called with one
argument, the name of the session to open, as in `topen
name-of-existing-session`. A list of existing sessions can be found via
the `tmux ls` command.