Print warning messages in Vim plugin update script
The update.sh script we use to update the Vim plugins in the ~/.janus
directory will now print a message to the user when it is done.
This message implores the user to commit their changes to git as soon
as
possible, and warns them about undefined behavior if they run the
update.sh script again without first committing their changes.

This should hopefully stop people (me) from causing problems with the
submodule system used to import Vim plugins.