Vim: strip trailing whitespace on save for code
This commit adds a function which strips the trailing whitespace from
a file. It first maps this function to the command `:Strip`, so that it
can be manually triggered by the user. It then adds an autocommand that
calls the function when code files (right now, a user-defined list of
filetypes I think I would want this done for) are saved.