Vim: use only openclc syntax checker for OpenCL
This sets the list of allowed syntax checker for Syntastic to include
only the openclc checker (which uses Apple's OpenCL compiler built into
OS X to compile and check the file for errors.)

You can trigger Syntastic to check a file for errors using the
`:SyntasticCheck` command. The errors will be highlighted, or can be
viewed as a list with the `:Errors` command.