Bash: Support Morgan Aldridge OS X tools directly
Instead of using an out-of-date copy of Morgan Aldridge's tools-osx
directory for things like our `trash` script, this commit allows us to
use a checkout of his tools-osx repo directory.

We first `rm` our out-of-date `trash` script from `./bin/trash`. Then,
in ./.bash_profile, we add a line which adds the 'src' directory of our
tools-osx checkout to the PATH.

While tools-osx nominally comes with an install script, this script
copies all of the scripts contained in that repo to /usr/local/bin.
I would rather not have these scripts installed into a global
directory,
especially not one managed by Homebrew, so it's easier to just add
their
location to the PATH (and to compile the one MacRuby script manually.)