Bash: change path to tools-osx; add diagnostic msg
We change the path of tools-osx from
'~/Documents/code/mac-scripts/tools-osx/src' to
'~/Documents/code/tools-osx/bin'. This keeps it out of our mac-scripts
directory, which is already linked in ~/. It also makes sure to add the
compiled bin directory, not the src directory.

Our ~/.bash_profile script will now display a message if it doesn't
find
tools-osx in the expected spot on terminal login. This is the only set
of tools .bash_profile will audibly complain about if missing, simply
because it's so useful, yet easy to forget to download and compile it
when installing a new system.