Bash: add local npm bins to $PATH
This commit changes the way the $PATH variable is constructed in
~/.bash_profile.

It adds a check to see if npm is installed and, if so, adds a relative
path entry to $PATH for local npm modules' binaries. These are the
binaries from npm modules installed without using the '-g' (global)
flag, and exist in ./node_modules/.bin/ when present.

This commit also re-arranges and cleans up the commands used in
constructing the $PATH.