webcl-test: default to re-launching WebKit
Previously, the '--dev' flag indicated that WebKit should be
re-launched when it exits, and to open the Web Inspector on WebKit
launch.

This commit inverts that behavior, so that, by default, WebKit is
re-launched and the inspector opened. The new '--once' flag is used to
turn off this behavior, and replicate the old default behavior.