webcl-test: display Notification Center messages
If the utility `terminal-notifier` is installed, webcl-test will
display a Mac OS X Notification Center notification every time it
re-launches WebKit, informing the user of the re-launch and how to stop
it. Clicking the message activates the Terminal.

Only one of these notifications exists in Notification Center at a time
(the old one is cleared when a new one is added,) and all notifications
are cleared when webcl-test exits.