Fix webcl-test bug opening Inspector
The `webcl-test` script was not properly opening the Web Inspector
panel
on WebKit.app launch.

The reason has to do with commit #57441bd, where we changed the default
action to be 'dev' mode (open WebKit in a loop, open the Web Inspector
on launch) instead of 'launch once' mode (open WebKit once, then finish
when it exits, and don't show the web inspector.) The AppleScript
launching WebKit.app was checking to see if dev mode was true before
opening the Inspector, when it really should check if 'launch once'
mode
is not set before opening the Inspector.