Bash: make `fgrep` case-insensitive; unalias `fg`
`fgrep` is a standard, system-installed alias to `grep -F`. We
previously added the alias `fg` to be case-insensitive `fgrep`.
However, in practice, I never remembered this, and just ran `fgrep -i`
all the time.

We remove the (underused) `fg` alias from Bash, and overwrite the
existing `fgrep` alias with our own, which runs `grep -Fi` (grep with a
fixed non-regex, fixed string, and ignore case differences).