Bash:set `xargs` to expect input is null separated
Usually `xargs` expects input to be whitespace separated, so that "one
two three" will run `xargs`'s given command three times. This commit
aliases `xargs` to have the "input is null separated" flag set, so that
"one two three" will be one command argument, but "one\0two\0three"
will be three. This is usually safer (if we can ensure that the program
generating input to `xargs` uses nulls), and can be used with things
like `find -print0` and the like.

This may break some scripts that depend on the default `xargs`
behavior, so watch out for that, and back out this commit if we notice
anything weird.