Bash: add `psgrep` command, to search `ps` output
The `psgrep` command will search the output of `ps` for a string, and
display the results on-screen in a human-friendly way.

Contrast this with `pgrep`, which searches running process names for a
given string, but only outputs the results as a list of pids, which are
hard for a human to make sense of.