.js: support dotjs-universal by moving configs
We move from the seemingly abandoned main dotjs project to the actively
updated dotjs-universal project. One of the changes is that it uses
separate script and style directories. To support this, we create
subdirectories to ~/.js for scripts and styles, and move our config
files to there. We also add the (currently blank) default config files
to each.