Remove our (broken) powerline config
Our powrline config settings in ~/.config/powerline currently is broken,
to the point that powerline will fail to run when it's in place. This
means that we have to rename/remove the directory before we can start
powerline.

This commit removes the config, so as to stop breakage. We can
re-configure powerline and some point in the future, in a working way.