Add default, placeholder email filters
The real filters are kept out of this (public) repo for security reasons. However, if they don't exist, git will complain. This commit adds three pass-through filters so that new checkouts of this repo won't have a broken git. They are meant to be replaced ASAP with the real filters.