Bash: fix some bugs in how we set $PROMPT_DIR_TRIM
We now only set $PWD_MAX_WIDTH if it's not already set. This allows us
to set it on a per-machine basis in ~/.config/bash/shell_local .

In the Ubuntu version of Bash, if we try to extract a substring of
greater length than the source string, the result in an empty string
(instead of the whole source string, as in OS X.) We now limit the size
of the string length to extract to the size of the source string.

Finally, we set a minimum size of 1 for $PROMPT_DIR_TRIM, since a size
of 0 means "don't trim any directories from PWD".