tmux: don't corrupt ssh-agent tunnel on reconnect
This enables the option in tmux's config to update $SSH_AUTH_SOCK
whenever we create or reconnect to a tmux session. This *should* fix the
problem where, when re-connecting to an existing tmux session with a new
SSH agent forwarding socket (as happens when we hibernate or restart OS
X), SSH agent forwarding wouldn't work unless we created a new shell.

I should note that this is only lightly tested and may not be sufficient
to solve the problem (though initial testing is positive.)