Bash: replace both `tnew`/`topen` aliases with `tm`
Instead of having the `tnew` alias to create a new session, and `topen`
to reconnect to an existing session, the `tm` alias will always connect
to a session named $USER. If the session does not exist, it will create
it and connect to it.

This mimics the behavior of .bash_profile creating/reconnecting to the
session $USER when we log in.