Bash: set/link $SSH_AUTH_SOCK to constant for tmux
This commit allows SSH agent forwarding to work in tmux, even on
disconnect and reconnect. It is taken from
http://unix.stackexchange.com/questions/75681/why-do-i-have-to-re-set-env-vars-in-tmux-when-i-re-attach

The problem is that when ssh disconnects and reconnects, a new SSH agent
forwarding socket is used, with a random name of the form /tmp/ssh-*
. Howver, in the existing tmux session, $SSH_AUTH_SOCK is still set to
the old socket name.

This commit adds logic to ~/.bash_profile which will symlink the current
$SSH_AUTH_SOCK file to ~/.ssh/agent_auth_sock, then export the value of
$SSH_AUTH_SOCK as '~/.ssh/agent_auth_sock'. This means that the SSH
agent forwarding socket will be in a constant location. Thus, even old
tmux sessions will still be able to use the current socket since its
location is the same.