Bash: set npm global dir to ~/.npm_global on Linux
On Ubuntu, `npm` will default to using a global directory under
`/usr/local`. This means that any global operations (e.g. `npm install
-g ...`) require `sudo`. This sucks just on its face, but what's worse
is that it can cause all sorts of permission errors if you need to
install global and local packages as a dependency of something (stuff
like packages in your current local package directory being owned by
root.)

This commit adds an environmental variable that tells npm to use
`~/.npm_global` as its global directory. It only does this on Linux.
This allows global npm operations to not need root. The downside is that
global operations are no longer truly global; they're only global for
the current user.